Support hierarchical scheduling
For this, a VCPUThread is added.

Also a scheduler class is added, to remove the scheduler_data hack and
to provide common functionality: next_start_time.
This is used to properly mark VCPUThreads as ready.
Actual hierarchical scheduling might work differently though, since
child scheduler might not be able to tell parents whether they have
threads that are ready or not.

The TextLog format got better alignment.