Track context stack in CPU
Previously the call chain would represent the context stack, starting
with the kernel scheduler, going to the child scheduler, the thread and
the thread would call execute_thread on the core.
Now the CPU tracks the call stack, so it is in control of returning
execution to the parent. Also the core calls execute() on a thread, not
the other way around.

The new approach is detailed in HACKING.

This is an invasive change, touching all schedsi files, but with this
change we can easily implement multi-core and external interrupts.
It also makes more sense to have the CPU at the bottom of the call-chain
and not the top.

This does complicate thread implementation slightly and schedulers
significantly, because they are interruptible in a sense. They are at
fixed points, but in preparation for modeling scheduler execution times
and multi-core scheduling without static partitioning the schedulers
need to deal with some synchronization.
This is implemented via RCU. As a bonus, since the RCU emulation is
thread-safe, the schedulers should also be thread-safe now.

The binary log also emulates the context stack of the CPU when replaying.
To achieve this the logs gained an init_core event, which the binary log
uses to obtain the initial context stack containing the kernel scheduler.
Direction recording has changed so that replaying the stack is possible.

With this commit, threads becoming ready are added to the ready queue of
the scheduler as they become ready and removed from it when they are
not, so the order schedulers select threads may change.