Make tid parameter for thread optional
It will be automatically set to the number of threads in the module,
which should be fine for most cases.

However, the current method of adding multiple threads in one go causes
the thread to take a tid without the module having added the previous
threads in the array.
Adding one thread at a time is inconvenient, we need a builder or
something.