Add MLFQ scheduler
This implements a multi-level feedback queue as described in "Operating
Systems: Three Easy Pieces" by Remzi H. Arpaci-Dusseau and Andrea C.
Arpaci-Dusseau.

So far documentation is missing, it's not included in the example or
tests (albeit it seems to work so far) and only implements approach #1,
meaning priority boosting and better accounting is missing.

While this is not a weighted round robin, this is what we were actually
looking for, so remove the TODO.