Make Scheduler._start_schedule return more information
This replaced the bool `removed` with the queue the thread is in now and
the index into `ready_threads` where the thread was.
If the queue is still `ready_threads` then the last thread can be
accessed via the index, if the list has not been modified beyond
appending (which is all Scheulder does).
Otherwise the last thread can be found of the end of either other
queue.