Add cpu.Cores as VCPUs for modules
This is so that num_threads is correct.

We use num_threads for automatic thread naming (see Thread.__init__),
but we add VCPUs and Cores after other threads.
This commit includes a hack that adds a None placeholder.
This will break once multi-core is supported, since more than one
VCPU/Core will be added.

Instead we should just remove the assumption that 0 is a special thread.