Avoid calling finish() on kernel scheduler
A thread locks the is_running lock when the execute() coroutine is
started and releases it when finish() is called.
The kernel scheduler is always kept on the stack. This means that the
execute() coroutine is only started once and then kept running.
Calling finish() on the thread will release the lock, but since the
coroutine is never restarted it expects the lock to be locked, which is
no longer the case.
Solution: avoid calling finish() on the kernel scheduler.