[BROKEN] Smarter plot.py parallelization
Instead of making a job per module-stats, make one per thread-stats.
This increases the amount of jobs, decreases the time each job takes and
results in better balancing.

Individual add_subplots are used in favor of creating all of them at
once, as this function is slow and thus benefits from the
parallelization.

However, plotting to the same figure from multiple processes is not
supported by matplotlib, so this code is broken.
It is faster though.