Make Thread._execute(current_time, 0) less awkward
Some threads need to update some execution stats without actually having
any time to spend executing. Calling Thread._execute() is a bit awkward
for this case, because the function may yield values, which we don't
expect it to do for this case.
Thread._execute(current_time, 0) is thusly moved to Thread._get_ready(),
which never yields.