Avoid passing the scheduler thread to the scheduler
This was done for a single specific purpose: to communicate the thread's
background time, which corresponds to the scheduler's previous thread's
execution time, which can be used by scheduling algorithms.

We avoid passing the thread by instead passing a communication channel,
to which the scheduler thread writes the background time and the
scheduler reads it.