Implement local-timer scheduling
Of course this has a lot of churn in cpu.py, but also the loggers must
be adapted.
Especially the context stack emulation by the BinaryLog, which must now
deal with splits and appending multiple contexts.

The single-timer scheduling approach is still available through the
cpu._KernelTimerOnlyStatus.
The approach to use is configured via the `local_timer_scheduling`
parameter to the Core.

A timer-request with "None" is now allowed. This is used for the
_KernelTimerOnlyStatus to ignore the request from non-kernel schedulers.

All threads now allow continuous execution without calling finish().