Complete incomplete statistics
Incomplete statistics were ignored as of
5b1c5c38f930f691bb12074e3dd049876ab3db26. This does not work well with
the local-timer scheduling, as most statistics remain incomplete.
Instead of leaving them out, we pass the current time to the statistics
gathering function so that we can complete incomplete statistics for
that purpose.

We can see that some context switching accounting differs between the
single-timer and local-timer scheduling strategies. This should not be
and needs to be investigated.