Include time_slice in Scheduler._sched_loop return value
Other schedulers were modifying Scheduler.time_slice, which was not
introduced for this purpose.
Instead schedulers can now return the desired time slice from
_sched_loop(). self.time_slice can be returned if the scheduler does not
choose time slices itself.

The SchedulerAddonBase is extended to allow modification of the
time slice.
With this we don't need the PenaltyMLFQ anymore:
The per-queue time-slices are integrated directly into MLFQ.
The generic PenaltySchedulerAddon is changed to intercept the
time-slices and use them to track penalties instead.

The change to MLFQ also allows the local-timer hierarchical scheduling
to leverage per-queue time-slices without the PenaltySchedulerAddon.