Always charge initiator with context switching
Previously we would charge the scheduler thread only when switching *to*
something. When switching *from* something, it would not be charged.

This seemed consistent, because we were just always charging the current
context (before the context switch), but it doesn't make sense to charge
worker threads with context switching time.

As a bonus, this also removes a hack when recording context switching on
the kernel.