Keep scheduler threads suspended until they have threads ready
This is achieved through another list to let the scheduler communicate
with the scheduler thread, like prev_run_time/bg_time.
The scheduler writes the time the next thread is ready before idling.
When the scheduler thread sees the scheduler making an idle request, the
channel is checked and ready_time is set accordingly, so that the
scheduler thread is being let sleeping the appropriate amount of time.