Switch -1 to None for undetermined times
Undetermined times or 'infinity' were encoded as -1. Special care had to
be taken on comparisons.

While special care still has to be taken on comparisons, they will fail
loudly if compared against an undetermined time.

This also switches the `units` parameter for HierarchyBuilderThread from
"None" for "finish thread after start" to -1, essentially reversing the
switch for this case.