Fix bg_time accounting for module switches
Module switches would count towards bg_times when the VCPU is appended
first and the VCPU switches context in the next step, but not when a
whole context chain is appended, which includes a module switch.

Since the intension is to not count for context switches initiated by
the module itself, doing the module switch after appending the VCPU no
longer counts towards bg_times of the same module.