Fix CFS thread insertion order
It seems the assumption was made that an inserted thread has the minimum
vruntime. At the same time, there was code in place to search the whole
ready queue for the insertion place.
The assumption is bad and now the search is actually useful.

Further, threads are now inserted at the end, so that insertion order is
preserved.

And a little comment fix in the tests.