Use a Fraction to represent time values
We used Pythons primitive number-type, which will become a float when
necessary. This introduced floating-point inaccuracies.
Since all we really need are fractions, a Time-type is introduced.
This was fractions.Fraction at first, which came with a big slowdown
compared to the floats before.
It is now a gmpy2.mpq, which performs significantly better.