Fix AddonSchedulerBase mro
We need the AddonSchedulerBase to be inserted *just* before the
Scheduler.
This was only the case for schedulers that directly inherit from
Scheduler.

SJF and RR for instance had their base classes after the
AddonSchedulerBase, resulting in misbehavior when an Addon blocks a
scheduling decision, due to a missing _start_scheduling() call in their
base classes.