Always add thread-switching cost when not switching to VCPUThread
Previously, when resuming a context chain with a different module at the
top, we'd only pay the module switching cost. Now we also pay thread
switching cost.
If the destination is a VCPUThread, the thread will switch to the
scheduler thread anyway, which is why we don't pay additional thread
switch cost when switching to a VCPUThread.