Change MODULE_CTXSW_COST behavior
Change between threads from different modules now always costs
MODULE_CTXSW_COST.
Changing between threads within the same module costs THREAD_CTXSW_COST,
except when we're changing to a VCPUThread. This does not incur any
cost, because an actual implementation would likely not implement the
VCPU as a regular thread that switches to the child, because that's two
context switches. Instead the scheduler would directly switch to the
child thread with a single switch.